Rework Blade Rush Calculations.
The current approach for Blade Rush suggestions doesn't match up with
existing guides/Ravenholdt recommandations. It is also using it's own
implementation of cooldown tracking which leads to some calculations
being wrong.

Changes:

* Use UpdateSpellUsableEvent instead of relying on an module specific
Cooldown tracking implementation

* Pull out suggestions into their own module so that the parsing logic
and the suggestion logic are decoupled.

* Take into account the fight type and energy levels at the time of the
ability going on cooldowns.(1)

(1): These are naive approaches so they aren't fully accurate. E.g: If
at anypoint the fight has a single add it will be considered AoE; if
the energy level is above the threshold at the time of the cast then
the late cast isn't punished.